fix: refactor annotations and edit
