+++ streamlibs/operations/edit/state.js
@@ -1,0 +1,10 @@
+export default function createEditState() {
+  return {
+    draggedPanelBlock: null,
+    draggedMainBlock: null,
+    dropPlaceholder: null,
+    currentDropContainer: null,
+    originalFigmaBlocks: [],
+    originalDABlocks: [],
+  };
+}
